+++ src/engine/wavefront/WavefrontMetricsRecorder.cpp
@@ -58,8 +58,7 @@ namespace engine::wavefront::detail {
     ++m_metrics.batching.batches;
     m_metrics.timings.sampleGenerationWorkerSeconds += result.sampleGenerationWorkerSeconds;
     m_metrics.timings.recordIntegratorBatch(result.integratorBatchWorkerSeconds,
-                                            result.batchMetrics.intersectionWorkerSeconds,
-                                            result.batchMetrics.shadingWorkerSeconds);
+                                            result.batchMetrics);
     m_metrics.batching.samplesSubmitted += result.sampleCount;
     m_metrics.batching.addIntegratorMetrics(result.batchMetrics);
     m_metrics.batching.maxBatchSize =