+++ src/render/Integrator.cpp
@@ -30,6 +30,10 @@ namespace render {
     stoppedAfterDepth = 0;
     intersectionWorkerSeconds = 0.0;
     shadingWorkerSeconds = 0.0;
+    pathSetupWorkerSeconds = 0.0;
+    frontierBookkeepingWorkerSeconds = 0.0;
+    progressSnapshotWorkerSeconds = 0.0;
+    convergenceTestWorkerSeconds = 0.0;
   }
 
   void IntegratorBatchMetrics::recordActiveDepth(std::uint64_t activeSamples) {